Use python script to download Antarctic ice shelves
The `download_cartopy_data.sh` script now generates a small
python script in the temp directory where `feature_download.py`
is downloaded.  This script first adds `antarctic_ice_shelves_polys`
to the list of features to download before performing the download
with a direct call to the `download_features` function.